1. 수락시 게시글 삭제하는데 알림에는 해당 알림 존재 -> 수락 시 알림 후 자동으로 삭제           onPressed: () async {
                        DocumentSnapshot postId = await FirebaseFirestore.instance
                            .collection('helpActions')
                            .doc(documentId)
                            .get();

                        String deletePostId = postId.get('post_id');
                        DocumentSnapshot postSnapshot = await FirebaseFirestore.instance
                            .collection('naver_posts') // 컬렉션 나주에 바꾸기
                            .doc(deletePostId)
                            .get();

                        if(postSnapshot.exists) {
                          DateTime now = DateTime.now();

                          HapticFeedback.lightImpact();
                          await _HelperCount(documentId);
                          await _updateTime(documentId, now);
                          await _respondToActions(documentId,
                              'accepted'); // ChatActions : null -< accept

                          await _deletePost(documentId); // 수락시 게시글 삭제          ////////출시할때 수정
                          _deleteNotification(documentId); // 수락시 알림 내용 삭제

                          Navigator.of(context).pop();

                          // ScaffoldMessenger 호출을 여기서 안전하게 실행
                          WidgetsBinding.instance?.addPostFrameCallback((_) {
                            if (mounted) {
                              ScaffoldMessenger.of(context).showSnackBar(
                                SnackBar(
                                  content: Text(
                                    '해당 요청이 수락되었습니다.',
                                    textAlign: TextAlign.center,
                                  ),
                                  duration: Duration(seconds: 1),
                                ),
                              );
                            }
                          });

                          Navigator.pushReplacement(
                            context,
                            MaterialPageRoute(builder: (context) =>
                                AllUsersScreen()), // 로그인 화면으로 이동
                          );
                        }

                        else if (!postSnapshot.exists){
                          ScaffoldMessenger.of(context).showSnackBar(
                            SnackBar(
                              content: Text(
                                '해당 게시글은 이미 거래가 성사되었습니다',
                                textAlign: TextAlign.center,
                              ),
                              duration: Duration(seconds: 2),
                            ),
                          );
                          _deleteNotification(documentId);
                          Navigator.of(context).pop(); // 다이얼로그 닫기
                          return;
                        }
2. 애플워치 로고 - o

3.
게시물 컬렉션 로그인한 도메인으로 설정 -> naverUserHelpStatus, naver_posts 수정해
도메인 naver.com 이면 -> naver_com으로 변경
naverUserHelpStatus -> UserHelpStatus
naver_posts -> 도메인에서 . 부분을 _ 로 교체

4.
파이어베이스 데이터베이스 보안설정 , Firebase Storage의 보안 설정 완료

240924 한일
신청 마감 프린트 - o
충대 계정 5개 생성 - o
원킥 계정 - o
지은이 계정 - o
ppt수정(교수님 발표) - o
2시 교수님 미팅- o
발표 대본 정리 - o
5시 hrd 발표- o
크몽 디자인 결제(카드뉴스) - o
빈 게시글 텍스트 통일 -o
배달비 헬퍼비로 수정 - o
카카오맵 지도 수정  - o

<예정>
가운데에서 위치 끄렁오기

거래 완료시 전체 광고 / 데베에 거래 완료된 횟수 측정하는거 넣기(보류)

인스타 광고시 피드 실제 대화사례 등 실사례위주

twilio 넣기
포스터나 팜플렛 제작시 -> 도와주시 신청과 받는 학생의 장점  / 최소주문을 강조  -인스타용도 조사ㅎ보기

#마지막으로 해야할일
개인정보 보호규정
앱개발 마케팅 영상
다이소, 편의점도 가능하다는 것을 강조
포스터나 팜플렛 제작시 -> 도와주시 신청과 받는 학생의 장점  / 최소주문을 강조
배달업체에게도 줘보자 -> 한명 정해보기(이건 미정)
개인정보 수집문?

구체적으로 알려주자 - 신뢰감, 내신상?
2. 다른 학교 추가시
    1. 게시판 수정 - ui
    2. schoolScores 컬렉션 도메인 문서 생성
    3. 학교랭킹 추가 - SchoolRanking.dart,
    4. 초기 중심 위치 설정 (현재 한밭대) - 게시글 작성시
    5. 회원가입시 학교 도메인 추가 - SchoolEmailDialog.dart
    6. 로고사진 스토리지에 넣기
    7. url 파이어스토어에 저장 - updateSchoolLogos()
    8. 바텀 게시판 이동 해당 학교 board 클래스작성
    9. AllUsersScreen.dart, SchoolRanking.dart, Profile.dart, 주문내역에 새로운 학교 import
    10. function index.js -> 대학별 알림 추가
    11. 회원탈퇴시 삭제할때 게시판 이름
    12. uiboard 에 boardCount 수정

3. 테스트 메일 삭제 - gmail.com

+++ lib/Alarm/AlarmUi.dart
@@ -19,7 +19,8 @@ class _NotificationScreenState extends State<AlarmUi> {
   late final Alarm alarm;  // NaverAlarm 클래스의 인스턴스를 선언합니다.
   late Stream<List<DocumentSnapshot>> notificationsStream; // 알림을 스트림으로 받아오는 변수를 선언합니다.
   bool isDeleteMode = false; // 삭제 모드 활성화 변수
-
+  String botton_domain = ""; // 사용자의 도메인을 저장할 변수
+  String collection_domain = "";
 
 
   @override
@@ -27,6 +28,10 @@ class _NotificationScreenState extends State<AlarmUi> {
     super.initState();
 
     final currentUserEmail = FirebaseAuth.instance.currentUser?.email ?? '';
+
+    botton_domain = currentUserEmail.split('@').last.toLowerCase();
+    collection_domain = botton_domain.replaceAll('.','_');
+
 
     alarm = Alarm(currentUserEmail, () {
       if (mounted) {
@@ -119,7 +124,7 @@ class _NotificationScreenState extends State<AlarmUi> {
     String deletePostId = postId.get('post_id');
 
     FirebaseFirestore.instance
-        .collection('naver_posts')
+        .collection(collection_domain)
         .doc(deletePostId)
         .delete()
         .then((_) => print("Document successfully deleted"))
@@ -600,7 +605,6 @@ class _NotificationScreenState extends State<AlarmUi> {
   //수락 확정 다이어로그
 
   void showNicknameConfirmationDialog(BuildContext context, String nickname, String documentId) {
-    final rootContext = context;
     showDialog(
       context: context,
       builder: (BuildContext context) {
@@ -662,18 +666,31 @@ class _NotificationScreenState extends State<AlarmUi> {
                     child: TextButton(
                       onPressed: () async {
 
-                        DateTime now = DateTime.now();
+                        DocumentSnapshot postId = await FirebaseFirestore.instance
+                            .collection('helpActions')
+                            .doc(documentId)
+                            .get();
 
-                        HapticFeedback.lightImpact();
-                        await _HelperCount(documentId);
-                        await _updateTime(documentId, now);
-                        await _respondToActions(documentId, 'accepted'); // ChatActions : null -< accept
+                        String deletePostId = postId.get('post_id');
+                        DocumentSnapshot postSnapshot = await FirebaseFirestore.instance
+                            .collection(collection_domain)
+                            .doc(deletePostId)
+                            .get();
 
-                        await _deletePost(documentId); // 수락시 게시글 삭제          ////////출시할때 수정
-                        _deleteNotification(documentId); // 수락시 알림 내용 삭제
+                        if(postSnapshot.exists) {
+                          DateTime now = DateTime.now();
+
+                          HapticFeedback.lightImpact();
+                          await _HelperCount(documentId);
+                          await _updateTime(documentId, now);
+                          await _respondToActions(documentId,
+                              'accepted'); // ChatActions : null -< accept
+
+                          await _deletePost(documentId); // 수락시 게시글 삭제          ////////출시할때 수정
+                          _deleteNotification(documentId); // 수락시 알림 내용 삭제
 
 
-                        Navigator.of(context).pop();
+                          Navigator.of(context).pop();
 
                           // ScaffoldMessenger 호출을 여기서 안전하게 실행
                           WidgetsBinding.instance?.addPostFrameCallback((_) {
@@ -692,11 +709,26 @@ class _NotificationScreenState extends State<AlarmUi> {
 
                           Navigator.pushReplacement(
                             context,
-                            MaterialPageRoute(builder: (context) => AllUsersScreen()), // 로그인 화면으로 이동
+                            MaterialPageRoute(builder: (context) =>
+                                AllUsersScreen()), // 로그인 화면으로 이동
                           );
+                        }
 
+                        else if (!postSnapshot.exists){
+                          ScaffoldMessenger.of(context).showSnackBar(
+                            SnackBar(
+                              content: Text(
+                                '해당 게시글은 이미 거래가 성사되었습니다',
+                                textAlign: TextAlign.center,
+                              ),
+                              duration: Duration(seconds: 2),
+                            ),
+                          );
+                          _deleteNotification(documentId);
+                          Navigator.of(context).pop(); // 다이얼로그 닫기
+                          return;
+                        }
                       },
-
                       style: TextButton.styleFrom(
                         padding: EdgeInsets.zero, // 여백을 제거하여 Divider와 붙도록 설정
                       ),
@@ -1023,7 +1055,7 @@ class _NotificationScreenState extends State<AlarmUi> {
                                                                   ScaffoldMessenger.of(context).showSnackBar(
                                                                       SnackBar(
                                                                         content: Text(
-                                                                          "해당 삭제되었습니다.", textAlign: TextAlign.center,),
+                                                                          "해당 알림이 삭제되었습니다.", textAlign: TextAlign.center,),
                                                                         duration: Duration(seconds: 1),
                                                                       ),
                                                                   );
